MAINT: separate data init and functions
